--- shard/master/fastapi/functions.py
@@ -1,8 +1,0 @@
-from db import SessionLocal
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
