+++ shard/master/fastapi/models.py
@@ -1,8 +1,9 @@
 from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import Column, Integer, String
 
 Base = declarative_base()
 
-# Модель User
+# User model
 class User(Base):
     __tablename__ = "users"
